+++ src/rex.setup/demo/rex.setup_demo/static/js/lib/index.js
@@ -1,0 +1,8 @@
+
+require('./index.less');
+var $ = require('jquery');
+
+$(function () {
+    $('body').html('<h1>Welcome to <tt>rex.setup_demo</tt>!</h1>');
+});
+