+++ src/rex.setup/src/rex/setup/commonjs/webpack.config.js
@@ -1,0 +1,5 @@
+'use strict';
+
+var configureWebpack = require('./index').configureWebpack;
+
+module.exports = configureWebpack({});